+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ v1.0.1
 Added
 -----
 * Changelog
+* ``interp_bounds`` formatoption for the ``plot2d`` plot method (see the
+  `docs <https://psyplot.readthedocs.io/projects/psy-simple/en/latest/api/psy_simple.plotters.html#psy_simple.plotters.Simple2DPlotter.interp_bounds>`__)
 
 Changed
 -------